load team members after adding
